BoneOrthoSystem/sub_system(old)/s2_agent/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
requests==2.32.3
reportlab==4.2.5
python-pptx==1.0.2
requests==2.32.3
openai>=1.53.0
httpx==0.27.2
PyPDF2
pdfplumber
python-docx
openpyxl
pypdfium2
